+++ tests/test_inject.py
@@ -1,3 +1,5 @@
+import pytest  # type: ignore
+
 from helpers import run_pipx_cli
 
 
@@ -11,9 +13,23 @@ def test_spec(pipx_temp_env, capsys):
     assert not run_pipx_cli(["inject", "pycowsay", "pylint==2.3.1"])
 
 
-def test_include_apps(pipx_temp_env, capsys):
-    assert not run_pipx_cli(["install", "pycowsay"])
-    assert run_pipx_cli(["inject", "pycowsay", "black", "--include-deps"])
+@pytest.mark.parametrize("with_suffix,", [(False,), (True,)])
+def test_include_apps(pipx_temp_env, capsys, with_suffix):
+    install_args = []
+    suffix = ""
+
+    if with_suffix:
+        suffix = "_x"
+        install_args = [f"--suffix={suffix}"]
+
+    assert not run_pipx_cli(["install", "pycowsay", *install_args])
+    assert run_pipx_cli(["inject", f"pycowsay{suffix}", "black", "--include-deps"])
+
+    if suffix:
+        assert run_pipx_cli(
+            ["inject", "pycowsay", "black", "--include-deps", "--include-apps"]
+        )
+
     assert not run_pipx_cli(
-        ["inject", "pycowsay", "black", "--include-deps", "--include-apps"]
+        ["inject", f"pycowsay{suffix}", "black", "--include-deps", "--include-apps"]
     )