+++ tests/test_upgrade.py
@@ -5,3 +5,12 @@ def test_upgrade(pipx_temp_env, capsys):
     assert run_pipx_cli(["upgrade", "pycowsay"])
     assert not run_pipx_cli(["install", "pycowsay"])
     assert not run_pipx_cli(["upgrade", "pycowsay"])
+
+
+def test_upgrade_suffix(pipx_temp_env, capsys):
+    name = "pycowsay"
+    suffix = "_a"
+
+    assert not run_pipx_cli(["install", name, f"--suffix={suffix}"])
+    assert run_pipx_cli(["upgrade", f"{name}"])
+    assert not run_pipx_cli(["upgrade", f"{name}{suffix}"])